feat(infra/web): TS entry point + MCPConfigManager; webpack.config.ts
Goal 3 (JS -> TS) for the web frontend:
- src/index.js -> src/index.ts (entry point; webpack entry now ./src/index.ts)
- src/components/MCPConfigManager.js -> .ts (typed: McpConfig, ConfigurationStatus,
  NotificationType; global Window.overlayCompanionApp augmentation)
- Add src/guacamole-common-js.d.ts type shim (library ships without types) so the
  remaining guacamole-dependent components can be converted later under strict mode.
- webpack.config.js -> webpack.config.ts (removed; ts-loader + .ts resolve added
  in prior commit) with ESM-safe __dirname resolution for ts-node config loading.
- tsconfig.json: remove noEmit (ts-loader needs to emit transpiled output to webpack).
- devDependencies: add ts-node + tsx (webpack-cli TS config loader).

Verified: tsc --noEmit passes; webpack --mode production builds successfully
(index.ts + app.js bundle into main.js + index.html). The remaining browser
components (app.js, GuacamoleClient, KasmVNCClient, OverlaySystem, StatusMonitor)
stay as .js for now (tsconfig allowJs:true) and are staged for conversion with
browser runtime testing in a follow-up.
