+++ infra/web/package.json
@@ -19,6 +19,8 @@
     "html-webpack-plugin": "^5.5.3",
     "style-loader": "^3.3.3",
     "ts-loader": "^9.5.1",
+    "ts-node": "^10.9.2",
+    "tsx": "^4.23.1",
     "typescript": "^5.7.0",
     "webpack": "^5.89.0",
     "webpack-cli": "^5.1.4",
