--- infra/web/webpack.config.js
@@ -1,85 +1,0 @@
-const path = require('path');
-const HtmlWebpackPlugin = require('html-webpack-plugin');
-const CopyWebpackPlugin = require('copy-webpack-plugin');
-
-module.exports = (env, argv) => {
-  const isProduction = argv.mode === 'production';
-
-  return {
-    entry: './src/index.js',
-    output: {
-      path: path.resolve(__dirname, 'dist'),
-      filename: isProduction ? '[name].[contenthash].js' : '[name].js',
-      clean: true,
-      publicPath: '/'
-    },
-    module: {
-      rules: [
-        {
-          test: /\.css$/i,
-          use: ['style-loader', 'css-loader']
-        },
-        {
-          test: /\.(png|svg|jpg|jpeg|gif)$/i,
-          type: 'asset/resource'
-        },
-        {
-          test: /\.(woff|woff2|eot|ttf|otf)$/i,
-          type: 'asset/resource'
-        }
-      ]
-    },
-    plugins: [
-      new HtmlWebpackPlugin({
-        template: './src/index.html',
-        title: 'Overlay Companion MCP',
-        inject: 'body'
-      }),
-      new CopyWebpackPlugin({
-        patterns: [
-          {
-            from: 'src/assets',
-            to: 'assets',
-            noErrorOnMissing: true
-          }
-        ]
-      })
-    ],
-    devServer: {
-      static: {
-        directory: path.join(__dirname, 'dist')
-      },
-      compress: true,
-      port: 3000,
-      hot: true,
-      historyApiFallback: true,
-      proxy: {
-        '/api': 'http://localhost:8080',
-        '/health': 'http://localhost:8080',
-        '/mcp-config': 'http://localhost:8080',
-        '/ws': {
-          target: 'ws://localhost:8080',
-          ws: true
-        }
-      }
-    },
-    optimization: {
-      splitChunks: {
-        chunks: 'all',
-        cacheGroups: {
-          vendor: {
-            test: /[\\/]node_modules[\\/]/,
-            name: 'vendors',
-            chunks: 'all'
-          }
-        }
-      }
-    },
-    resolve: {
-      extensions: ['.js', '.json'],
-      alias: {
-        '@': path.resolve(__dirname, 'src')
-      }
-    }
-  };
-};
